% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{anti_mode}
\alias{anti_mode}
\title{anti_mode}
\usage{
anti_mode(data)
}
\arguments{
\item{data}{A vector or factor that you wish to find the anti_mode of}
}
\value{
The same data type as the data you passed it. The Anti_Mode of that data
}
\description{
This function returns the anti_mode(Least frequent element) of the data passed to it
}
\examples{
anti_mode(c("Alex","Ben","Alex","Alex",",Ben",",Tim",",Tom"))
}
