% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_annual_income}
\alias{gen_annual_income}
\title{gen_annual_income}
\usage{
gen_annual_income()
}
\value{
Numeric. The annual income
}
\description{
This function generates a random annual income that is at least 10000 and at most 620000 and always divisible by 500
}
\examples{
gen_annual_income()
}
