% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_birthday}
\alias{gen_birthday}
\title{gen_birthday}
\usage{
gen_birthday()
}
\value{
Date. A randomly selected Birthday
}
\description{
This function generates a random date that is at least 18 years ago and at max 65 yeas ago
}
\examples{
gen_birthday()
}
