% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_boy_name}
\alias{gen_boy_name}
\title{gen_boy_name}
\usage{
gen_boy_name()
}
\value{
String. Represents a name for a boy
}
\description{
This function generates a random boy name
}
\examples{
# Generate a random boy's name
gen_boy_name()
}
