% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_email_address}
\alias{gen_email_address}
\title{gen_email_address}
\usage{
gen_email_address(firstname, surname)
}
\arguments{
\item{firstname}{A String representing the first name of the person you want to generate an email for}

\item{surname}{A String representing the surname of the person you want to generate an email for}
}
\value{
String. Represents an Email Address
}
\description{
This function generates a fake email address using a first name and surname.
}
\examples{
# Generate an email address for Joe Miller
gen_email_address("Joe","Miller")
}
