% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_girl_name}
\alias{gen_girl_name}
\title{gen_girl_name}
\usage{
gen_girl_name()
}
\value{
String. Represents a name for a girl
}
\description{
This function generates a random girl name
}
\examples{
# Generate a random girl's name
gen_girl_name()
}
