% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_people}
\alias{gen_people}
\title{gen_people}
\usage{
gen_people(amount)
}
\arguments{
\item{amount}{The amount of people you want to generate}
}
\value{
Data Frame. A data frame with the specified amount of rows, each representing a person with the columns Name, Surname, Sex, Email, Birthday, and Annual_Income
}
\description{
This function generates a data frame with information about lots of people
}
\examples{
# Generate information for 151 people
gen_people(151)
}
