% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_person}
\alias{gen_person}
\title{gen_person}
\usage{
gen_person()
}
\value{
Data frame. It represents a person with the columns Name,Surname,Sex,Email,Birthday and, Annual income
}
\description{
This function generates a random person
}
\examples{
# Generate a random person
gen_person()
}
