% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_sex}
\alias{gen_sex}
\title{gen_sex}
\usage{
gen_sex()
}
\value{
String. It represents a sex("male" or "female")
}
\description{
This function generates a random sex(male or female)
}
\examples{
gen_sex()
}
