% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_surname}
\alias{gen_surname}
\title{gen_surname}
\usage{
gen_surname()
}
\value{
String. Represents a common surname
}
\description{
This function generates a random surname
}
\examples{
# Generate a random surname
gen_surname()
}
