% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(data)
}
\arguments{
\item{data}{A vector or factor that you wish to find the mode of}
}
\value{
The same data type as the data you passed it. The Mode of that data
}
\description{
This function returns the mode of the data passed to it
}
\examples{
mode(c("Alex","Ben","Alex","Alex",",Ben",",Tim",",Tom"))
}
